[FIRRTL][LowerCHIRRTL] Fix enable inference for aggregates
Sequential memory read ports have enable inference when the address is
something with a name.  We missed that using the subfield of an
aggregate does not count as something with a name, and should tie the
enable of the port to 1.

This adds a more complicated check to turn on enable inference for ports
which verifies that the value used is a port, wire, reg, regreset, or
node.