[PrepareForEmission] Spill wires for largish expressions in concat.
This adds another heuristic to shouldSpillWire, this time to check if
the number of operands in a single expression inside a concat. If this
exceeds a configurable limit, a wire is spilled. This is to work
around Verilator #3405.

Large expressions inside a concat is worst case O(n^2) in a certain
Verilator optimization, and can effectively hang Verilator on large
designs. Verilator 4.224+ works around this by having a hard limit on
its recursion. Here we break large expressions inside concats
according to a configurable limit to work around the same issue.
See https://github.com/verilator/verilator/issues/3405.