[FIRRTL][AddSeqMemPorts] Don't error if there are still MemOps
The `AddSeqMemPorts` pass has a check during execution that it doesn't
see any MemOps, since they should have been converted to behavioural
implementations or blackboxed.  This was a very simple check added just
to catch any phase-ordering problems in the FIRRTL pipeline.  We have
since revised the lowering of memories to all some memories to be
handled by the HW memory generator, and so some mem ops can still be
around and it isn't an error.