Revert "[FIRRTL] Use symbols for metadata emission (#3096)"
This reverts commit 832de8c7ccf2cdb0ab0284c07a5aea874d6de018.

This PR reverts the recent improvement to metadata emission, which used
symbols to ensure that the metadata reflected changes in the hierarchy.
It also modified the CreateSiFiveMetadata to pass to be able to run
before BlackBoxMemory. This just involved adding one extra annotation to
the list of known BlackBoxReader annotations.