[FIRRTL] Produce pre-extract seqmem metadata
The metadata for seq_mems.json needs to be created before extract
instances runs, since it needs the original paths to the memory
operations. Moving it before the inliner makes us output identical
instance paths as SFC, which the downstream tools seem to be able to
handle.

It turns out that we don't want to have the final instance paths
represented in the metadata.

Eventually, we will be getting rid of this particular piece of metadata,
and replacing it with OM.