[FSM] Move unreachable state canonicalization to separate pass
As pointed out in #3020, erasing other operations in a canonicalizer may occasionally trip up the infrastructure.
As such, unreachable state canonicalization has been moved to a separate pass.
This is done, instead of moving it to be a canonicalization of `StateOp`, to avoid DFS being run for _every_ state in the FSM.