[Calyx] Initial Calyx to HW conversion. (#3012)
This adds initial support for converting Calyx to the "core" CIRCT
dialects: HW, Comb, Seq, and SV. This is done in a fairly direct way
using the dialect conversion framework. This is direct in the sense
that a Calyx reg is mapped to a Seq reg, a Calyx assign is mapped to a
SV assign, a Calyx std_add is mapped to a Comb add, etc.

I tried to port big chunks of the Calyx primitives over at once, when
the conversion could be templated and shared. There are still some ops
missing conversions.

I also deviate slightly in the behavior of pipelined ops, like
multiplication. This may lead to slight cycle-by-cycle differences
with the Calyx standard library implementations, but should still
produce functionally identical behavior.

The implementation makes use of lots of SV wires and assigns, in order
to stay simple and in-line with the form of the Calyx IR. These are
massively cleaned up by canonicalization, which can be run after the
conversion.