[FIRRTL] Stop Blocking Optimizations on GCT Views
Change Grand Central (GCT) views feature to stop putting a don't
touch (optimization barrier) on signals that it is tapping.  Instead,
teach IMCP and canonicalization to not remove things which are part of a
GCT view.

Admittedly, this is a giant hack and we should rip this out in the
future in favor of a less special-cased approach.  E.g., this would be
better handled with access modifiers on symbols (public/private,
read/write) or dedicated operations (read tap, write tap).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>