[FIRRTL] Honor symbols in RemoveUnusedPorts
Adjust the `RemoveUnusedPorts` pass such that it takes into
consideration whether a port has a symbol or not.

Output ports with a symbol now always lead to the creation of a local
wire with that symbol to act as a stand-in for the original port for
anyone holding a reference to that symbol. This is probably not the
right thing to do in all circumstances, because users of the symbol
might actually rely on it being a port, but we have no solid mechanism
to detect this. Therefore we now somewhat recklessly assume that what
matters is that the symbol can be replaced with a valid name during
Verilog emission later, and just emit this wire.

Input ports with a symbol cannot be removed, since we don't know what
value to drive onto a replacement wire that would act as a stand-in for
the port. This might warrant some sort of `firrtl.gct.tap` op.