[FIRRTL] Set name and file location for Comb memory wrapper module (#3150)
This commit sets the Combinational memory wrapper module name and file location,
instead of the generic "FIRRTLMem_0_2_0_8_16_1_1_1_0_1_0_a" formatted name and
default file location.
Use the name associated with firrtl.mem to derive the wrapper module name, after
appending it with _combMem.
This ensures the correct module prefix is used for the wrapper module. (For
example SiFive prefix for DUT modules and not otherwise).
For memories outside the DUT, set the output file directory to the TestBench
directory if available.
This change requires to add the MemOp to the FirMemory, such that the memory
dedup can be done in LowerToHW.
Use a map memoryNameMap to get the correct HWModuleGeneratedOp names for deduped
memories.