[CalyxToFSM] Initial Calyx-to-FSM control schedule lowering
This commit represents a lowering pass for converting a Calyx control schedule (seq/if/while/enable) into an FSM representation. The FSM is embedded within the Calyx component, and references both the group symbols and cell SSA values.

The lowering method is a fairly straight forward conversion, which leaves plenty of canonicalization opportunities to remove redundant states in the generated FSM. However, doing it as presented in this PR lends itself to some very clean code (which I prefer, rather than prematurely optimizing during lowering) as well as meaningful naming of states (which i think is fairly important).

Marked as draft due to mixing in changes to the FSM dialect which needs to be discussed. Once things are settled, FSM dialect changes and FSMGraph changes will be factored out into a separate PR.

Things to be discussed:
- MachineOp is *no more `IsolatedFromAbove`*; this is a requirement for the FSM to be able to reference the symbols and SSA values defined within the component. If we find that the presented style of lowering is compelling, but we still want an `IsolatedFromAbove` FSM machine, one option could be to introduced a new `EmbeddedMachineOp` to allow for this kind of behaviour.
- I added a `fsm.ref` (not sure about the name) to represent _a reference to some symbol_. The semantics of a reference is implementation defined, and merely serves to convey the semantics of associating a symbol reference with a given state. `fsm.ref` is call-like and can provide arguments to references, as well as return values.