Temporarily stop printing version info in generated files
In b7e1ad378 we started adding a line to print the version of CIRCT
which generated a file.  This code did not discriminate the type of file
we were emitting and would create illegal content for some file formats.
For now, this just reverts the version printing code.  In the future, we
should only emit this string in Verilog files, not every emitted file.