[FIRRTL] simplify by using tokenizePath, NFCI, drop sigmap options
Support emitting instance path XMR's if the annotation
inputted to the subcircuit asks for them.
(tokenizePath makes handling this easy and clean, thanks!)

In the full flow we still will emit the `module.name` XMR's
(our updated mappings will be in simple `~circuit|module>name` form)
as that's simpler and doesn't require DUT/prefix reasoning, but with
this working that should be straightforward when if/when we want to
implement that too.

Vectors/bundles are maybe a little sketchy presently,
preserve current behavior for now, which assumes the remote aggregates
have been broken apart and given names with underscores.
One wart is that multiple forces targeting same value may be emitted if
local target is a vector, for example, as demonstrated in tests
currently.

NFC(I).

Also, drop markDut/prefix options as they're unused.  They live on in
git history should we need to call on them again in our time of need.