[firtool] Only blackbox SRAMs when `-repl-seq-mem` is specified
Due some hasty changes when we realized that MemToRegOfVec transform
should not transform all memories, we started blackboxing seq mems which
we shouldn't be.  We need to make sure that the LowerMemory pass is
gated on the `repl-seq-mem` flag in firtool, which will allow SRAMs to
be lowered to generated memory modules.