Fix website generation for pass documentation
When we generate a markdown file, and it is not included in any other
markdown file, the website generator will assume that the MD file should
be a top-level page.  Since the MD files generated for passes are
fragments, we end up with some bad top-level pages and untitled entries
in the side bar.