[FIRRTL] Use tap wires for GCT data taps on module ports
Instead of directly referring to module ports, this commit changes the
GCT data taps pass to create an auxiliary "tap wire" inside the module.
The wire is connected such that it "observes" the current value of the
port (input or output), and all symbol references generated by GCT are
with respsect to this wire.

This has two distinct advantages:

1.  It removes a major source of symbols on module ports, which block
    port optimizations almost entirely.
2.  It offers a "receptacle" for IMConstProp to propagate constants
    into. This can lead to the tap wire being disconnected from the port
    if the port is only driven to a constant externally, which allows
    for the port to be removed.