[FIRRTL] Change Grand Central Views to use Wire Taps and Sink Constants (#3167)
Modify Grand Central (GCT) Views to use read-only wire taps instead of
directly tapping the targeted signals.  This change enables GCT Views to
not block optimizations due to tap points no longer getting inner
symbols (DontTouchAnnotations).

Improve Grand Central (CGCT) views handling to use constants instead of
XMRs when the XMR is tapping a wire/node that is driven by a constant.

E.g., GCT will now generate an XMR like the following where previously
it would XMR to a wire driven by a constant:

    assign Foo.bar.baz = 1'h1;

When this logic is used, also remove the symbol on the tapped wire.  This
effectively causes the constant tap to be a candidate for later DCE and
removes a source of dead wires in the output Verilog.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>