[FSM] Remove IsolatedFromAbove trait from `fsm.machine`
As discussed in #3138, the IsolatedFromAbove trait is removed from the `fsm.machine` to allow it to be embedded within other operations, wherein it is possible to reference SSA values and symbols defined within an outer scope.

This change is fairly motivated by using `fsm.machine` inside a Calyx component. If it is, in the future, found that this change is prohibitive in writing generic/performant transformations on `fsm.machine`s, then FSM should take presendence, with the `IsolatedFromAbove` trait being re-introduced and the Calyx pass re-evaluated.