[CalyxToFSM] Add FSM materialization pass
In materializing a Calyx control FSM we:
* Gather up all of the `calyx.group`s referenced in the FSM
* Create an in- and output for each group, representing `done` (FSM input) and `go` (FSM output) signals
* Iterate over all states in the FSM and
  * replace `calyx.enable`s with an `fsm.output` bitvector asserting the corresponding `go` outputs
  * Extend the guards of all transitions from the state with a conjunction of the `done` signals of the enabled groups in the given state.

The materialization does not consider how to materialize static groups. However, this should be fairly easy to implement/hook into when iterating over states and detecting `calyx.group`s with `static` attributes.

This pass was added under `CalyxToFSM` with the alternative being to make it a `Calyx` pass. Please let me know what you think is the most suitable.