[FIRRTL] Sink Constants in GCT Data Taps
Change Grand Central (GCT) Data Taps to sink trivial constants produced
by IMCP or canonicalization of "dead wire" taps into the actual data
tap.  The symbol on the "dead wire" tap is then removed to enable it to
be removed.

This causes some problems with existing tests that use invalid or
constant values.  Tests are updated to prevent this via either
DontTouchAnnotations or verbatim expressions (which block
optimizations).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>