[FIRRTL] Use tap wires for GCT data taps (#3192)
Instead of directly referring to module ports, this commit changes the
GCT data taps pass to create an auxiliary "tap wire" inside the module.
The wire is connected such that it "observes" the current value of the
port (input or output), and all symbol references generated by GCT are
with respsect to this wire.

This has two distinct advantages:

1.  It removes a major source of symbols on module ports, which block
    port optimizations almost entirely.
2.  It offers a "receptacle" for IMConstProp to propagate constants
    into. This can lead to the tap wire being disconnected from the port
    if the port is only driven to a constant externally, which allows
    for the port to be removed.

Change GCT Data Taps to use a tap wire _for all_ taps, not just ports.
This is done to not block optimizations on tapped wires/nodes/registers
and to provide a constant receptacle for a constant value driven to
these wires.

Change Grand Central (GCT) Data Taps to sink trivial constants produced
by IMCP or canonicalization of "dead wire" taps into the actual data
tap.  The symbol on the "dead wire" tap is then removed to enable it to
be removed.

This causes some problems with existing tests that use invalid or
constant values.  Tests are updated to prevent this via either
DontTouchAnnotations or verbatim expressions (which block
optimizations).

Co-authored-by: Fabian Schuiki <fabian@schuiki.ch>
Co-authored-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>