[ExportVerilog] Emit a version string only to verilog files (#3174)
This fixes an issue that version string is emitted to non-verilog files
such as json, metadata files #3169. This commit adds an additional flag
`isVerilog` to `FileInfo` so that we can emit version strings
only to verilog files in SharedEmitterState::collectOpsForFile.
Currently, files are regarded as verilog if output file names has ".sv" suffix.