[MSFT] Track instance hierarchy through module parameters (#3229)
The MSFT dialect will be specializing modules based on a _per-instance_ basis. In practice, we expect this to be rather sparse. Rather than elaborating the entire design, we've chosen to select implementations based on the instance hierarchy path. So we need a way to know said path. Tracking it through module parameters fills that need.

As a first step, this PR adds a pervasive `__INST_HIER` module parameter during lowering and appends the instance name to it in every module instance.