[FIRRTL] Add some more functionality to `InnerSymbol`
This adds a getter, setter, and rmeover for the `inner_sym` attribute to
the `InnerSymbolOpInterface`.  This folows the same conention as
`Symbol` and names the functions `getName()` and `setName()`.  This is
pretty confusing since most things using this interface have another
attribute actually called `name`.

This fixes up most places in the FIRRTL passes that used to manually
look for the `inner_sym` attribute.  We still have a few places that
deal with symbols on ports.