[SCFToCalyx, Calyx, CalyxToHW] Add support for signed div and rem (#3238)
Adds support both at the front- and back-end level to ensure that all signed `arith` division operators work end-to-end.
Notably, this also changes the semantics of the `DivLibPipe` operator in that it no longer provides both a quotient and remainder through division. This change is motivated by the fact that:
- `arith.div` operations only provide quotients
- `comb.div` operations only provide quotients

As such, this change aligns the calyx division operators to better fit within an MLIR/CIRCT up/downstream context.