[FIRRTL] Make all firrtl.hierpath end in a Module
Change FIRRTL's Hierarhical Path to require that it ends in a module.
This is a semantic simplification that means a Hierarchical Path only
encodes a path and does not encode what it points to.  This is done for
two reasons:

First: this enables NLAs to exist without requiring a symbol on their
leaf target.  This avoids a problem where this symbol, which is an
optimization barrier equivalent to a DontTouchAnnotation, would block
optimizations unnecessarily.  Additionally, symbols apply to the whole
aggregate as opposed to individual fields which resulted in a loss of
optimization barrier specificity when using NLAs.

Second: this enables Hierarchical Paths to be shared amongst multiple
"circt.nonlocal" annotations (NLAs).  NLA verification is expensive and
this has the potential to speed up this verification by avoiding
repeated work repeatedly verifying the existence of the same symbols.

Note: this commit makes no changes to Hierarchical Path generation in
parsing/scattering that works towards the second reason.  This will be
added in later commits.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>