squash! [FIRRTL] Make all firrtl.hierpath end in a Module
Update LowerMemory to understand and update the new format for
hierarchical paths that end in a module and not an inner ref.  This uses
existing symbols if they are available and doesn't generate unnecessary
symbols on the "ext" memory module.

Due to inability to know if a hierarchical path has users, this
generates a new hierarchical path for each lowered memory.  If two
memories share a hierarchical path, then they will share _one_ new
hierarchical path.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>