[FIRRTL] Properly parse printf assertions with no message
The assertion parsing framework would parse the label from
`cover:my-label:` as the message.  This was a mistake in part of the
logic to handle missing labels in such cases as "cover:my message",
where it would always treat the label as a message.  Now, it uses the
colon to disambiguate the two cases.

As a follow up we should check if it is actually legal to have a missing
label.