[FIRRTL] Properly parse printf assertions with no message (#3286)
The assertion parsing framework would parse the label from
`cover:my-label:` as the message.  This was a mistake in part of the
logic to handle missing labels in such cases as `cover:my message`,
where it would always use label as a message if the message was empty.
Now, it uses the colon to disambiguate the two cases.

As a follow up we should check if it is actually legal to have a missing
label.