Improve documentation of attributes
This PR aims to improve the generated documentation for some attribute
defintions and constraints. This PR tries to fix some issues where
missing tablegen fields created empty markdown headings with no content.
This focus of this PR is just getting anything in these fields, without
much focus on creating high quality content. Somewhat confusingly, the
`summary` field of attribute constraints is used as the "name" of the
constraint in most locations.