Improve documentation of attributes (#3296)
This PR aims to improve the generated documentation for some attribute
definitions and constraints. Some missing tablegen fields create empty
markdown headings with no content.  The focus of this change is just
getting anything in these fields, without much focus on creating high
quality content. Somewhat confusingly, the `summary` field of attribute
constraints is used as the "name" of the constraint in many locations.