[LowerToHW] Fix bug with unmasked RW memory lowering (#3282)
This commit fixes an issue with lowering unmasked RW memories.
If a memory has a single bit mask, then the mask bit can be removed and
combined with the write enable. But if the memory is a ReadWrite memory,
then the enable port is used for both read and write enable. For such a
ReadWrite memory we cannot combine the write mask with the enable, it has
to be combined with the Write Mode. This ensures that if the mask bit is
disabled, then the write should be disabled, but not the read.