[FIRRTL] Update Dedup to handle HierPathOp ending on Module (#3309)
This change updates the dedup pass to handle non-local annotations that
use a hierarchical path which ends at the parent module instead of the
annotation target. The main change needed is to track a list of
operations which reference the path operation, as multiple operations
can now share a single path. A follow up commit will change dedup to
produce the new style of path when it changes a local annotation into a
non-local one.