[FIRRTL] Update InjectDUTHierarchy for new NLAs (#3331)
Update the InjectDUTHierarchy pass to work for both old-style
NLAs (where an NLA that targets a port/component ends in a
hw::InnerRefAttr) and new-style NLAs (where an NLA that targets a
port/component ends in a module/FlatSymbolRefAttr).

With these changes, InjectDUTHierarchy will reuse the existing DUT path
for components which use it inside the wrapper.  A module target or port
target annotation on the DUT will use a clone of the original path (and
get a new symbol).  This is a more efficient approach as it doesn't
require any walk of the DUT internals to find users.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>