[LowerToHW] Add BackedgeBuilder, directly materialize inst inputs (#3323)
Add a `BackedgeBuilder` to `FIRRTLLowering` which can be used to
directly materialize values that violate def-before-use ordering. To
simplify usage, a new `setBackedgeLowering` function creates a backedge
with the desired type and memorizes it as the lowering for the original
value. A later call to `setLowering` for the original value will resolve
the backedge.

As a first use of the backedge builder, materialize the input values for
instances directly through a backedge if the port is trivially assigned
by a single `StrictConnectOp`. This does away with temporary wires in
most cases and is a useful preparation for handling non-FIRRTL types
during lowering.