[FIRRTL] Fix handling of new-style NLAs in ExtractInstances
Adapt `ExtractInstances` such that it can properly handle the new flavor
of NLAs, where non-local annotations would carry a symbol to a hierpath
that generally only points at their parent module.

The `ExtractInstances` pass had a few assumptions in its implementation
about hierpaths pointing at the annotation target op, and each anno
having a unique hierpath op associated with it. The implementation can
be simplified quite a bit once support for the old-style NLAs can be
ripped out.

This commit also adds explicit tests to check that the old *and* new
style of NLAs are working in parallel.

A nasty corner case still exists when extraction requires an NLA to be
re-routed because the instance has been extracted past the NLA's
original root. If the instance is multiply-extracted, one hierpath gets
turned into multiple hierpaths, which will require special handling of
*everything* with that hierpath. This commit detects the common case
which we can handle and ensures that the hierpath is re-rooted in place,
and produces a warning about a potential breakage of hierpath references
in the corner case.