[LowerToHW] Generalize handling of non-FIRRTL ops
Adjust `FIRRTLLowering::handleUnloweredOp` to apply a consistent
lowering to operations from dialects other than FIRRTL. This used to be
a very specific snippet of code that allowed `hw.output` and the
`unrealized_conversion_cast` inserted after block args to be passed
through.

With this change, `handleUnloweredOp` will check if an operation comes
from outside of FIRRTL. If it does we replace all operands with a
potentially lowered value, and record the results as identity lowerings.
As a result, we no longer have to have special treatment of the
`unrealized_conversion_cast` inserted after block args.

This is a first step towards foreign type support, and will cause
`LowerToHW` to do the right thing for operations from other dialects
that are present in the module (operand updating and recording of the op
as already lowered).

As an implementation detail, this requires adjusting
`getPossiblyInoutLoweredValue` to always consider block args as already
lowered, and making `setLowering` accept non-FIRRTL values. The latter
still asserts that the lowering is sane in case the pre-lowering value
has a FIRRTL type, but it admits non-FIRRTL pre-lowering values and
requires them to have a non-null lowering.