firtool: fix IR printing options --mlir-print-ir-{before,after}
Ensure the passes we use are registered before parsing the command line
options, so that options such as `--mlir-print-ir-before=` work.

Add test.

It would be better if we didn't have to worry about this,
("did we remember to register all the passes" / early loading)
but in the meantime these options are very helpful for debugging
large designs.

This may pull in more than strictly needed,
Debug binary went 135M -> 140M after this change.