[FIRRTL] EmitOMIR: Drop temporary HierPath's that may be invalid
Restore behavior for deleting the HierPath's added in this pass.

They're known to be dead and sometimes are invalid, causing
breakage during circuit verification.

While we cannot delete incoming HierPath's attached to operations in
input IR without a more careful analysis of all possible users, any NLA
explicitly created in this pass can be reasoned about and removed at the
end for cleaner (and valid) IR.

Fixes assertion failure on a real design.

The invalid NLA encountered is when the target is in the root module,
and so we end up making a one-element (InnerRef) NLA.