[FIRRTL] Use Minimal XMRs in GCT Views
Modify the Grand Central (GCT) Views pass to generate SV
Verbatim-encoded XMRs that are as short as possible.  Do this with a
naive lowest-common ancestor (LCA) computation the prunes the XMR path
to start at the GCT View parent module (the module where the mapping
module containing the XMRs is instantiated).

Previously, the full non-local annotation (NLA) hierarchical path would
be blindly used under the assumption that this path was what the user
wanted.  This was problematic as the Chisel API will always generate an
absolute path starting from the FIRRTL circuit's main module instead of
a path starting from where the GCT API is used.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>