[ESI] Inline ChannelBufferOptions into ChannelBuffer op.
StructAttr was finally removed from upstream MLIR in
d883a02a7c2bb89000d0685749f062c9206ac40c. ESI had a peculiar usage of
StructAttr, ChannelBufferOptions, that did not migrate cleanly to the
new way of defining Attrs with multiple parameters. The issue was that
ChannelBufferOptions acted as a dictionary of attributes for
the ChannelBuffer op, but this adds an extra layer of key-value pair
nesting under the new scheme.

Since the two properties of ChannelBufferOptions were orthogonal, it was
cleaner to just define these as regular, scalar attributes directly on
ChannelBuffer.

This commit inlines these two properties of ChannelBufferOptions
directly onto ChannelBuffer and updates all the ESI code to directly
access the attributes on the ChannelBuffer op. It slightly changes the
assembly format in the case where no attributes are specified, where
instead of writing a pair of empty braces (`{ }`), the braces are
omitted altogether.

This ultimately simplified the custom assembly format parser, however,
since we can just directly use the `parseOptionalAttrDict()` method now.