[FIRRTL] Use relative XMRs in GCT Data/Mem Taps
Change the Grand Central (GCT) Data/Mem Taps pass to use relative XMRs
for taps that involve non-local annotations (NLAs).  Previously, NLA
paths were used as-is and did not factor into existing XMR minimization
code.  Change this by adding NLA path information to the existing
InstanceOp array used by minimization code.

This is done to fix issues where a user creates a Data/Mem Tap at a
specific point in the hierarchy and they wind up with an XMR that goes
_above_ this point in the design.  This creates problems for designs
that expect the XMR to be compartmentalized, but where the Chisel API
produces NLA paths that are absolute.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>